asm: split instruction translation into dedicated files
